# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис сокращения ссылок на Fiber + GORM + Postgres",
        "title": "URL Shortener API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/r/{code}": {
            "get": {
                "tags": [
                    "links"
                ],
                "summary": "Редирект на оригинальный URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Короткий код",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    }
                }
            }
        },
        "/shorten": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Создать короткую ссылку",
                "parameters": [
                    {
                        "description": "JSON request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/stats/{code}": {
            "get": {
                "tags": [
                    "links"
                ],
                "summary": "Получить статистику по ссылке",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Короткий код",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ShortLink"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.CreateRequest": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "entity.CreateResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "shortUrl": {
                    "type": "string"
                }
            }
        },
        "entity.ShortLink": {
            "type": "object",
            "properties": {
                "clicks": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}